Fix puppet-syntax error UTF-8
Error: puppet-syntax: UTF-8: Could not parse for environment production:
invalid byte sequence

SELinux uses the `.pp` extension, which stands for "Policy Package".
This is great, except that puppet-syntax expects files that have the
`.pp` extension to be puppet manifests. This commit simply removes the
extension from this file. It's only used as the return value for a stub,
so it is safe for us to remove the extension for the sake of testing.