Remove certain files that should remain in puppet
There were a few files that had made their way into this module that
should stay in the puppet code base. This commit removes those files.

I had also missed a few spec files which we do want in the module. So
this commit pulls those in.

We are not using any of the helper methods form
spec/lib/puppet_spec/files.rb, so that file, and references to it, have
also been removed.